refactor: move session utils
