+++ packages/api-server/src/sessions/sessions.controller.ts
@@ -5,7 +5,6 @@ import {
   Body,
   Patch,
   UseGuards,
-  BadRequestException,
   ParseIntPipe,
   Param,
 } from '@nestjs/common';
@@ -37,7 +36,9 @@ export class SessionsController {
     createSessionDto.hostId = userId;
     const sessionFileIds = createSessionDto.sessionFileIds || [];
     delete createSessionDto.sessionFileIds;
-    await this.validateFileIds(userId, sessionFileIds);
+    for (const fileId of sessionFileIds) {
+      await this.filesService.getFileAsUser(fileId, userId);
+    }
     const session = await this.sessionsService.create(createSessionDto);
     for (const fileId of sessionFileIds) {
       await this.sessionFilesService.create(session.sessionId, fileId);
@@ -51,7 +52,7 @@ export class SessionsController {
     @Param('sessionId', ParseIntPipe) sessionId: number,
     @UserId(ParseIntPipe) userId: number,
   ) {
-    return await this.getSessionAsUser(sessionId, userId);
+    return await this.sessionsService.getSessionAsUser(sessionId, userId);
   }
 
   @Patch(':sessionId/info')
@@ -63,8 +64,13 @@ export class SessionsController {
   ) {
     const sessionFileIds = updateSessionDto.sessionFileIds || [];
     delete updateSessionDto.sessionFileIds;
-    const session = await this.getSessionAsUser(requestedSessionId, userId);
-    await this.validateFileIds(userId, sessionFileIds);
+    const session = await this.sessionsService.getSessionAsUser(
+      requestedSessionId,
+      userId,
+    );
+    for (const fileId of sessionFileIds) {
+      await this.filesService.getFileAsUser(fileId, userId);
+    }
     for (const sessionFile of await this.sessionFilesService.findAllBySessionId(
       session.sessionId,
     )) {
@@ -79,21 +85,5 @@ export class SessionsController {
     }
     await this.sessionsService.update(session.sessionId, updateSessionDto);
     return null;
-  }
-
-  async getSessionAsUser(sessionId: number, userId: number | null = null) {
-    const session = await this.sessionsService.findOne(sessionId);
-    if (!session || (userId && session.hostId !== userId)) {
-      throw new BadRequestException(
-        `Session with ID: ${sessionId} does not exist or you do not have permission to access it.`,
-      );
-    }
-    return session;
-  }
-
-  async validateFileIds(userId: number, fileIds: number[]) {
-    for (const fileId of fileIds) {
-      await this.filesService.getFileAsUser(fileId, userId);
-    }
   }
 }
